# Add a couple of precompilation statements